# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Rest API",
        "description": "Laravel Rest API Swagger Documentation.",
        "contact": {
            "name": "RestAPI"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1": {},
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Brands List API",
                "description": "Brands List API",
                "operationId": "0c1a5d6b90006d89100119dd48d02d74",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome to Laravel Rest API"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/brands/{id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Brands Detail API",
                "description": "Brands Detail API",
                "operationId": "478503d73ae0cc58e25809e95e4e24ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome to Laravel Rest API"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Categories List API",
                "description": "Categories List API",
                "operationId": "9e975d8e2b6733988615c39b97de2eec",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome to Laravel Rest API"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Categories Detail API",
                "description": "Categories Detail API",
                "operationId": "508b6d6447014d67bc052ff71e595956",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome to Laravel Rest API"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}